Install auth

+++ database/migrations/2025_09_29_160731_create_event_table.php
@@ -7,27 +7,28 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up(): void
     {
-        Schema::create('events', function (Blueprint $table) {
-            $table->id('event_id'); // PK
-            $table->unsignedBigInteger('user_id'); // FK ke event_creators
+        if (! Schema::hasTable('events')) {
+            Schema::create('events', function (Blueprint $table) {
+                $table->id('event_id');
+                $table->unsignedBigInteger('events_creators_id'); // FK ke event_creators.id (creator)
 
-            $table->string('event_name');
-            $table->text('event_description');
-            $table->string('event_location');
-            $table->dateTime('event_date');
-            $table->integer('event_capacity');
-            $table->timestamps();
+                $table->string('event_name');
+                $table->text('event_description');
+                $table->string('event_location');
+                $table->dateTime('event_date');
+                $table->integer('event_capacity');
+                $table->timestamps();
 
-            // FK
-            $table->foreign('user_id')
-                ->references('user_id')
-                ->on('event_creators')
-                ->onDelete('cascade');
-        });
+                $table->foreign('events_creators_id')
+                    ->references('id')
+                    ->on('event_creators')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::dropIfExists('events');
+        // no-op
     }
 };
